backend/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-filter==24.2
pillow==10.3.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-cors-headers==4.3.1
pymupdf==1.24.5